Fixed the iterator types just like everywhere else

+++ ft_strncat.c
@@ -14,8 +14,8 @@
 
 char	*ft_strncat(char *s1, const char *s2, size_t n)
 {
-	int i;
-	int j;
+	size_t i;
+	size_t j;
 
 	if (!(s1))
 		return (NULL);
